Admin panel backend mcp file (cloudstorage system file) deletion

+++ UT4MasterServer.Services/Scoped/CloudstorageService.cs
@@ -21,6 +21,23 @@ public sealed class CloudStorageService
 	private readonly IMongoCollection<CloudFile> cloudStorageCollection;
 	private readonly ILogger<CloudStorageService> logger;
 
+	private static readonly string[] commonSystemFileFilenames = new string[]
+	{
+		// yes, 3 of these are misspelled by epic
+		"UnrealTournamentOnlineSettings.json",
+		"UnrealTournmentMCPAnnouncement.json",
+		"UnrealTournmentMCPGameRulesets.json",
+		"UnrealTournmentMCPStorage.json",
+		"UTMCPPlaylists.json"
+	};
+	private static readonly string[] commonUserFileFilenames = new string[]
+	{
+		// old players might also have "user_profile_1", but it is not used for anything anymore
+		"user_profile_2",
+		"user_progression_1",
+		"stats.json"
+	};
+
 	public CloudStorageService(DatabaseContext dbContext, ILogger<CloudStorageService> logger)
 	{
 		cloudStorageCollection = dbContext.Database.GetCollection<CloudFile>("cloudstorage");
@@ -32,16 +49,10 @@ public sealed class CloudStorageService
 		// get a list of already stored system files
 		var stored = await ListFilesAsync(EpicID.Empty);
 
-		// get a list of default system files
-		var files = Directory.EnumerateFiles("CloudstorageSystemfiles");
-
 		// ensure that all default files exist in db
-		foreach (var file in files)
+		foreach (var filename in commonSystemFileFilenames)
 		{
-			// get just the filename part of file path
-			var filename = Path.GetFileName(file);
-			if (filename == null)
-				continue;
+			var file = Path.Combine("CloudstorageSystemfiles", filename);
 
 			if (stored.Any(x => x.Filename == filename))
 			{
@@ -100,9 +111,16 @@ public sealed class CloudStorageService
 		return await cursor.ToListAsync();
 	}
 
-	public async Task DeleteFileAsync(EpicID accountID, string filename)
+	public async Task<bool?> DeleteFileAsync(EpicID accountID, string filename)
 	{
-		await cloudStorageCollection.DeleteOneAsync(GetFilter(accountID, filename));
+		if (accountID.IsEmpty && !IsCommonSystemFileFilename(filename))
+			return false;
+
+		var result = await cloudStorageCollection.DeleteOneAsync(GetFilter(accountID, filename));
+		if (!result.IsAcknowledged)
+			return null;
+
+		return result.DeletedCount > 0;
 	}
 
 	public async Task<int?> RemoveFilesByAccountAsync(EpicID accountID)
@@ -116,8 +134,12 @@ public sealed class CloudStorageService
 
 	private static bool IsCommonUserFileFilename(string filename)
 	{
-		// old players might also have "user_profile_1", but it is not used for anything
-		return filename == "user_profile_2" || filename == "user_progression_1" || filename == "stats.json";
+		return commonUserFileFilenames.Contains(filename);
+	}
+
+	private static bool IsCommonSystemFileFilename(string filename)
+	{
+		return commonSystemFileFilenames.Contains(filename);
 	}
 
 	private static string CalcFileHash(byte[] data)
